+++ gemini-core/src/main/java/com/haroncode/gemini/connection/EventListenerConnection.kt
@@ -1,7 +1,6 @@
-package com.haroncode.gemini.routing
+package com.haroncode.gemini.connection
 
-import com.haroncode.gemini.connection.BaseConnectionRule
-import io.reactivex.Flowable
+import com.haroncode.gemini.routing.EventListener
 import io.reactivex.functions.Consumer
 import org.reactivestreams.Publisher
