+++ gemini-core/src/main/java/com/haroncode/gemini/connection/NavigationConnection.kt
@@ -1,20 +1,19 @@
-package com.haroncode.gemini.routing
+package com.haroncode.gemini.connection
 
-import com.haroncode.gemini.connection.BaseConnectionRule
+import com.haroncode.gemini.core.Store
 import com.haroncode.gemini.core.elements.StoreNavigator
 import io.reactivex.Flowable
 import io.reactivex.functions.Consumer
 import io.reactivex.rxkotlin.Flowables
-import org.reactivestreams.Publisher
 
 class NavigationConnection<State : Any, Event : Any>(
-    statePublisher: Publisher<State>,
-    eventPublisher: Publisher<Event>,
+    override val isRetain: Boolean,
+    store: Store<*, State, Event>,
     storeNavigator: StoreNavigator<State, Event>
 ) : BaseConnectionRule<Pair<State, Event>, Pair<State, Event>>(
     publisher = Flowables.combineLatest(
-        Flowable.fromPublisher(statePublisher),
-        Flowable.fromPublisher(eventPublisher)
+        Flowable.fromPublisher(store),
+        Flowable.fromPublisher(store.eventSource)
     ),
     consumer = Consumer { (state, event) -> storeNavigator.invoke(state, event) },
     transformer = { input -> input }
